Move test only source code to separate module called testutils
