+++ testutils/src/main/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<manifest
+    package="com.microsoft.identity.internal.testutils">
+</manifest>
