+++ testutils/src/main/java/com/microsoft/identity/internal/testutils/authorities/B2CTestAuthority.java
@@ -1,9 +1,10 @@
-package com.microsoft.identity.common.internal.authorities;
+package com.microsoft.identity.internal.testutils.authorities;
 
+import com.microsoft.identity.common.internal.authorities.AzureActiveDirectoryB2CAuthority;
 import com.microsoft.identity.common.internal.logging.Logger;
 import com.microsoft.identity.common.internal.providers.microsoft.microsoftsts.MicrosoftStsOAuth2Configuration;
 import com.microsoft.identity.common.internal.providers.oauth2.OAuth2Strategy;
-import com.microsoft.identity.common.internal.providers.ropc.ResourceOwnerPasswordCredentialsTestStrategy;
+import com.microsoft.identity.internal.testutils.strategies.ResourceOwnerPasswordCredentialsTestStrategy;
 
 public class B2CTestAuthority extends AzureActiveDirectoryB2CAuthority {
